Addressed Mrinal's review comments and fixed panic
"odo service create" without any arguments was panicking.
